% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-sqlite.R
\name{create_sqlite}
\alias{create_sqlite}
\title{create_sqlite}
\usage{
create_sqlite(obofile, dbfile, name = "", date = "", url = "", ont2gene = NULL)
}
\arguments{
\item{obofile}{ontology store in an obo file}

\item{dbfile}{filename of the output sqlite file}

\item{name}{name of the ontology}

\item{date}{release date of the ontology}

\item{url}{reference url of the ontology}

\item{ont2gene}{ontology ID to gene ID mapping (optional)}
}
\description{
create sqlite file from an obo file
}
\author{
Guangchuang Yu \url{https://yulab-smu.top}
}
