% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-obo.R
\name{read.obo}
\alias{read.obo}
\title{read-obo}
\usage{
read.obo(file)
}
\arguments{
\item{file}{obo file}
}
\value{
a list of obo information, including term, relation, alias and synonym
}
\description{
parse obo file
}
\author{
Guangchuang Yu \url{https://yulab-smu.top}
}
